# docs/swagger.yaml
basePath: /
definitions:
...
  main.Service:
    properties:
      actif:
        type: boolean
      categorie:
        type: string
      created_at:
        type: string
      credits:
        type: integer
      description:
        type: string
      duree_minutes:
        type: integer
      id:
        type: integer
      provider_id:
        type: integer
      titre:
        type: string
      ville:
        type: string
    type: object
  main.ServiceRequest:
    properties:
      categorie:
        type: string
      credits:
        type: integer
      description:
        type: string
      duree_minutes:
        type: integer
      titre:
        type: string
      ville:
        type: string
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: API REST d'échange de compétences via crédits-temps.
  title: BarterSwap API
  version: "1.0"
paths:
...
  /api/services:
    get:
      description: Liste les services actifs avec filtres optionnels par catégorie,
        ville et recherche plein texte.
      parameters:
      - description: Filtrer par catégorie
        in: query
        name: categorie
        type: string
      - description: Filtrer par ville
        in: query
        name: ville
        type: string
      - description: Recherche plein texte
        in: query
        name: search
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/main.Service'
            type: array
      summary: Rechercher des services
      tags:
      - Services
    post:
      consumes:
      - application/json
      description: Publie une nouvelle annonce de service par le prestataire connecté
        (requiert la compétence correspondante).
      parameters:
      - description: ID du prestataire
        in: header
        name: X-User-ID
        required: true
        type: integer
      - description: Données du service
        in: body
        name: service
        required: true
        schema:
          $ref: '#/definitions/main.ServiceRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/main.Service'
        "400":
          description: Requête ou compétence invalide
          schema:
            type: string
        "403":
          description: Accès interdit
          schema:
            type: string
      summary: Créer une annonce de service
      tags:
      - Services
  /api/services/{id}:
    delete:
      description: Désactive un service (l'annonce ne sera plus visible ni disponible).
        Le prestataire connecté doit être le propriétaire du service et il ne doit
        pas y avoir d'échange actif.
      parameters:
      - description: ID du service
        in: path
        name: id
        required: true
        type: integer
      - description: ID du prestataire
        in: header
        name: X-User-ID
        required: true
        type: integer
      responses:
        "204":
          description: Désactivé avec succès
        "400":
          description: ID invalide
          schema:
            type: string
        "403":
          description: Accès interdit
          schema:
            type: string
        "404":
          description: Service non trouvé
          schema:
            type: string
        "409":
          description: Conflit (échange en cours)
          schema:
            type: string
      summary: Désactiver un service
      tags:
      - Services
    get:
      description: Récupère les informations d'un service actif par son ID.
      parameters:
      - description: ID du service
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/main.Service'
        "400":
          description: ID invalide
          schema:
            type: string
        "404":
          description: Service non trouvé ou inactif
          schema:
            type: string
      summary: Obtenir le détail d'un service
      tags:
      - Services
    put:
      consumes:
      - application/json
      description: Met à jour le titre, la description, la catégorie, la durée ou
        les crédits d'un service existant. Le prestataire connecté doit être le propriétaire
        du service.
      parameters:
      - description: ID du service
        in: path
        name: id
        required: true
        type: integer
      - description: ID du prestataire
        in: header
        name: X-User-ID
        required: true
        type: integer
      - description: Nouvelles données du service
        in: body
        name: service
        required: true
        schema:
          $ref: '#/definitions/main.ServiceRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/main.Service'
        "400":
          description: Requête invalide
          schema:
            type: string
        "403":
          description: Accès interdit
          schema:
            type: string
        "404":
          description: Service non trouvé
          schema:
            type: string
      summary: Modifier un service
      tags:
      - Services
...
swagger: "2.0"